Update CapturePage.xaml.cs
Bugfix: when the application was in fullscreen state, if for some reason he window is minimized the captured video flux was crashed.
Solution: Add an event handler on the Application window's event VisibilityChanged to pause the capture when the application is not visible and run again the capture when the application is visible.